refactor(bot/autoplay): break Cycles A + B residuals (partial #889) (#925)
## Summary

Partial fix for issue #889 — closes Cycles A and B of the 3 remaining
bot autoplay/queue cycles. Cycle C (`queueManipulation ↔ replenisher`)
deferred as follow-up.

**Madge cycle count:**
- Before: 3 runtime cycles + 1 type-only (4 total)
- After: 1 runtime cycle (`queueManipulation ↔ replenisher`) + 1
type-only (2 total)

**Cycle A — autoplayAudit hub** (CLOSED)
- Extracted `markAsAutoplayTrack` (and helpers) into
`packages/bot/src/utils/music/autoplay/queueMarkers.ts` (~49 LOC).
- `queueEditOps.ts` re-exports the symbols for non-breaking caller
migration.
- `diversitySelector.ts` imports directly from `queueMarkers.ts`
(back-edge removed).

**Cycle B — candidateCollector ↔ spotifyRecommender** (CLOSED)
- Extracted shared types + helpers into
`packages/bot/src/utils/music/autoplay/candidateContracts.ts` (~97 LOC).
- Both modules now import from `candidateContracts.ts` instead of each
other.

**Cycle C — replenisher utility back-imports** (DEFERRED)
- Agent attempt to extract `enrichWithAudioFeatures` +
`getTrackAudioFeatures` + `buildVcContributionWeights` +
`interleaveByArtist` (proposed `audioFeatures.ts` + `vcWeights.ts`)
broke 107 bot tests due to closure-captured state in
`queueManipulation`. Rolled back the broken extraction.
- Recommended follow-up: more careful extraction with proper handling of
shared internal helpers, or accept the cycle and document it in an ADR
per the issue's "two options" framing.

## Verified

- ✓ `npx madge --circular packages/bot/src --extensions ts,tsx`: 2
cycles (1 deferred type-only + 1 Cycle C residual)
- ✓ `npm test --workspace=packages/bot`: 2878 / 2878 tests pass

## Closes

Partial #889 (Cycles A + B). Cycle C remains open as the issue's
residual.
